完善service层方法。把实现写在service业务层。商品：增加分页逻辑，条件查询也支持分页。增加自定义异常类的处理，包括接口、枚举、异常类。

+++ src/main/java/com/szcf/vue_server/controller/ProductController.java
@@ -1,6 +1,7 @@
 package com.szcf.vue_server.controller;
 
 import com.szcf.vue_server.bean.Product;
+import com.szcf.vue_server.exception.BizException;
 import com.szcf.vue_server.service.impl.ProductServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -39,14 +40,26 @@ public class ProductController {
 
     // 搜索  多条件判断（sql）   支持模糊查询
     @RequestMapping("/search")
-    public String searchPro(String gName, String gCat){
-        return  productService.queryProByCondition(gName,gCat);
+    public String searchPro(String gName, String gCat,int currentPage,int pageSize){
+        System.out.println("搜索条件："+ gName+"  ---  "+ gCat);
+        return  productService.queryProByCondition(gName,gCat,currentPage,pageSize);
+    }
+
+    //  分页查询
+    @RequestMapping("/getProByPage")
+    public String getProByPage(int currentPage,int pageSize){
+        System.out.println("当前页："+ currentPage);
+        System.out.println("每页记录数："+ pageSize);
+        return productService.getByPage(currentPage,pageSize);
     }
 
     // 删除
     @RequestMapping("/deleteById/{gId}")
     public String deleteById(@PathVariable("gId") String gId){
         System.out.println("根据id删除："+ gId);
+        if(gId.equals("")){
+            throw new BizException("-1","传入id为空！");
+        }
         return productService.deleteById(gId);
     }
 
@@ -57,8 +70,6 @@ public class ProductController {
         return productService.updatePro(product);
 
     }
-
-
 
 
 }
